+++ downstream/job.go
@@ -2,6 +2,7 @@ package downstream
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -21,7 +22,7 @@ type JobParams struct {
 // Post makes a request to /v1/jobs/:job-name/:job-id with the job data.
 // The downstream service is expected to respond with a 202, so there is no
 // positive return value, only nil if the response was a 2xx status code.
-func (j *JobService) Post(name string, id *types.PrefixUUID, jp *JobParams) error {
+func (j *JobService) Post(ctx context.Context, name string, id *types.PrefixUUID, jp *JobParams) error {
 	if jp == nil || id == nil {
 		return errors.New("no job to post")
 	}
@@ -37,6 +38,7 @@ func (j *JobService) Post(name string, id *types.PrefixUUID, jp *JobParams) erro
 	if err != nil {
 		return err
 	}
+	req = req.WithContext(ctx)
 	var d struct{}
 	return j.Client.Do(req, &d)
 }