+++ models/archived_jobs/archived_jobs.go
@@ -20,8 +20,8 @@ var ErrNotFound = errors.New("archived_jobs: job not found")
 // Create an archived job with the given id, status, and attempts. Assumes that
 // the job already exists in the queued_jobs table; the `data` field is copied
 // from there. If the job does not exist, queued_jobs.ErrNotFound is returned.
-func Create(id types.PrefixUUID, name string, status newmodels.ArchivedJobStatus, attempt int16) (*newmodels.ArchivedJob, error) {
-	aj, err := newmodels.DB.CreateArchivedJob(context.Background(), newmodels.CreateArchivedJobParams{
+func Create(ctx context.Context, id types.PrefixUUID, name string, status newmodels.ArchivedJobStatus, attempt int16) (*newmodels.ArchivedJob, error) {
+	aj, err := newmodels.DB.CreateArchivedJob(ctx, newmodels.CreateArchivedJobParams{
 		ID:       id,
 		Name:     name,
 		Status:   status,