+++ models/queued_jobs/queued_jobs_test.go
@@ -1,6 +1,7 @@
 package queued_jobs
 
 import (
+	"context"
 	"fmt"
 
 	"github.com/kevinburke/go-types"
@@ -8,7 +9,7 @@ import (
 
 func ExampleDelete() {
 	id, _ := types.NewPrefixUUID("job_6740b44e-13b9-475d-af06-979627e0e0d6")
-	err := Delete(id)
+	err := Delete(context.TODO(), id)
 	// Returns an error, because we didn't insert rows into the database.
 	fmt.Println(err)
 }