+++ test/services/status_callback_test.go
@@ -1,6 +1,7 @@
 package services
 
 import (
+	"context"
 	"testing"
 
 	"github.com/kevinburke/rickover/models/archived_jobs"
@@ -14,9 +15,9 @@ import (
 func TestStatusCallbackInsertsArchivedRecordDeletesQueuedRecord(t *testing.T) {
 	defer test.TearDown(t)
 	qj := factory.CreateQueuedJob(t, factory.EmptyData)
-	err := services.HandleStatusCallback(qj.ID, "echo", newmodels.ArchivedJobStatusSucceeded, 7, true)
+	err := services.HandleStatusCallback(context.Background(), qj.ID, "echo", newmodels.ArchivedJobStatusSucceeded, 7, true)
 	test.AssertNotError(t, err, "")
-	_, err = queued_jobs.Get(qj.ID)
+	_, err = queued_jobs.Get(context.Background(), qj.ID)
 	test.AssertEquals(t, err, queued_jobs.ErrNotFound)
 	aj, err := archived_jobs.Get(qj.ID)
 	test.AssertNotError(t, err, "")
@@ -29,9 +30,9 @@ func TestStatusCallbackInsertsArchivedRecordDeletesQueuedRecord(t *testing.T) {
 func testStatusCallbackFailedInsertsArchivedRecord(t *testing.T) {
 	t.Parallel()
 	job, qj := factory.CreateUniqueQueuedJob(t, factory.EmptyData)
-	err := services.HandleStatusCallback(qj.ID, job.Name, newmodels.ArchivedJobStatusFailed, 1, true)
+	err := services.HandleStatusCallback(context.Background(), qj.ID, job.Name, newmodels.ArchivedJobStatusFailed, 1, true)
 	test.AssertNotError(t, err, "")
-	_, err = queued_jobs.Get(qj.ID)
+	_, err = queued_jobs.Get(context.Background(), qj.ID)
 	test.AssertEquals(t, err, queued_jobs.ErrNotFound)
 	aj, err := archived_jobs.Get(qj.ID)
 	test.AssertNotError(t, err, "")
@@ -41,9 +42,9 @@ func testStatusCallbackFailedInsertsArchivedRecord(t *testing.T) {
 func TestStatusCallbackFailedAtMostOnceInsertsArchivedRecord(t *testing.T) {
 	defer test.TearDown(t)
 	_, qj := factory.CreateAtMostOnceJob(t, factory.EmptyData)
-	err := services.HandleStatusCallback(qj.ID, "at-most-once", newmodels.ArchivedJobStatusFailed, 7, true)
+	err := services.HandleStatusCallback(context.Background(), qj.ID, "at-most-once", newmodels.ArchivedJobStatusFailed, 7, true)
 	test.AssertNotError(t, err, "")
-	_, err = queued_jobs.Get(qj.ID)
+	_, err = queued_jobs.Get(context.Background(), qj.ID)
 	test.AssertEquals(t, err, queued_jobs.ErrNotFound)
 	aj, err := archived_jobs.Get(qj.ID)
 	test.AssertNotError(t, err, "")
@@ -53,10 +54,10 @@ func TestStatusCallbackFailedAtMostOnceInsertsArchivedRecord(t *testing.T) {
 func testStatusCallbackFailedAtLeastOnceUpdatesQueuedRecord(t *testing.T) {
 	t.Parallel()
 	job, qj := factory.CreateUniqueQueuedJob(t, factory.EmptyData)
-	err := services.HandleStatusCallback(qj.ID, job.Name, newmodels.ArchivedJobStatusFailed, 7, true)
+	err := services.HandleStatusCallback(context.Background(), qj.ID, job.Name, newmodels.ArchivedJobStatusFailed, 7, true)
 	test.AssertNotError(t, err, "")
 
-	qj, err = queued_jobs.Get(qj.ID)
+	qj, err = queued_jobs.Get(context.Background(), qj.ID)
 	test.AssertNotError(t, err, "")
 	test.AssertEquals(t, qj.Attempts, int16(6))
 
@@ -67,10 +68,10 @@ func testStatusCallbackFailedAtLeastOnceUpdatesQueuedRecord(t *testing.T) {
 func testStatusCallbackFailedNotRetryableArchivesRecord(t *testing.T) {
 	t.Parallel()
 	qj := factory.CreateQJ(t)
-	err := services.HandleStatusCallback(qj.ID, qj.Name, newmodels.ArchivedJobStatusFailed, qj.Attempts, false)
+	err := services.HandleStatusCallback(context.Background(), qj.ID, qj.Name, newmodels.ArchivedJobStatusFailed, qj.Attempts, false)
 	test.AssertNotError(t, err, "inserting archived record")
 
-	_, err = queued_jobs.Get(qj.ID)
+	_, err = queued_jobs.Get(context.Background(), qj.ID)
 	test.AssertEquals(t, err, queued_jobs.ErrNotFound)
 	aj, err := archived_jobs.Get(qj.ID)
 	test.AssertNotError(t, err, "finding archived job")
@@ -83,6 +84,6 @@ func testStatusCallbackFailedNotRetryableArchivesRecord(t *testing.T) {
 func TestStatusCallbackFailedAtMostOnceArchivedRecordExists(t *testing.T) {
 	defer test.TearDown(t)
 	aj := factory.CreateArchivedJob(t, factory.EmptyData, newmodels.ArchivedJobStatusFailed)
-	err := services.HandleStatusCallback(aj.ID, aj.Name, newmodels.ArchivedJobStatusFailed, 1, true)
+	err := services.HandleStatusCallback(context.Background(), aj.ID, aj.Name, newmodels.ArchivedJobStatusFailed, 1, true)
 	test.AssertEquals(t, err, queued_jobs.ErrNotFound)
 }